refactor test cases and update mocking strategy
