+++ biz/controller/user.py
@@ -1,6 +1,7 @@
 import base64
 import os
-import requests
+from google.oauth2 import id_token
+from google.auth.transport import requests
 from typing import Optional
 
 from cryptography.exceptions import InvalidKey
@@ -32,12 +33,11 @@ def compare_password(hashed_password: str, password: str) -> bool:
 
 def validate_oauth_token(provider: str, access_token: str) -> bool:
     if provider == 'google':
-        url = f'https://www.googleapis.com/oauth2/v1/tokeninfo?access_token={access_token}'
-        response = requests.get(url)
-        if response.status_code != 200:
+        try:
+            id_info = id_token.verify_oauth2_token(access_token, requests.Request())
+            return 'email' in id_info
+        except ValueError:
             return False
-        data = response.json()
-        return data.get('email') is not None
     return False
 
 def signup_user_by_credential(email: str, password: str):
@@ -61,10 +61,11 @@ def signup_user_by_credential(email: str, password: str):
 def signup_user_by_oauth(provider: str, provider_account_id: str, email: str,
                          access_token: str, refresh_token: Optional[str], expires_at: Optional[int],
                          name: Optional[str], image: Optional[str]):
-    with get_session(write=True) as session:
-        if not validate_oauth_token(provider, access_token):
-            raise ResponseCode.InvalidAuth.create_error("Invalid or expired OAuth token.")
 
+    if not validate_oauth_token(provider, access_token):
+        raise ResponseCode.InvalidAuth.create_error("Invalid or expired OAuth token.")
+
+    with get_session(write=True) as session:
         account = Account.get_by_provider_and_provider_id(session, provider, provider_account_id)
         if account:
             raise ResponseCode.InvalidParam.create_error(f"{provider} account already registered.")
@@ -98,10 +99,9 @@ def login_user_with_credential(email: str, password: str):
 def login_user_by_oauth(provider: str, provider_account_id: str,
                         access_token: str, refresh_token: Optional[str],
                         expires_at: Optional[int]):
+    if not validate_oauth_token(provider, access_token):
+        raise ResponseCode.InvalidAuth.create_error("Invalid or expired OAuth token.")
     with get_session(write=True) as session:
-        if not validate_oauth_token(provider, access_token):
-            raise ResponseCode.InvalidAuth.create_error("Invalid or expired OAuth token.")
-
         account = Account.get_by_provider_and_provider_id(session, provider, provider_account_id)
         if not account:
             raise ResponseCode.InvalidParam.create_error(f"{provider} account not registered.")
